requirements.txt
alembic==1.7.1
auth==0.5.3
autopep8==1.5.7
blinker==1.4
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
ecdsa==0.17.0
eventlet==0.32.0
falcon==3.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==2.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jose==1.0.0
Mako==1.1.5
MarkupSafe==2.0.1
mongoengine==0.23.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pymongo==3.12.0
python-dotenv==0.19.0
python-jose==3.3.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
